WORLD LEVEL REFERENCE HELL YEAH

+++ include/data.h
@@ -65,7 +65,7 @@ struct TrainerMon
     u16 species;
     u16 heldItem;
     enum Ability ability;
-    u8 lvl;
+    u16 lvl;
     u8 ball;
     u8 friendship;
     u8 nature:5;
